move PanelContent to new layout/ directory
